Send the client's cameraId as part of config commands
When multiple users of the same instance are opened via client
APK, the camera works only on the first client connected. All
the other clients crash when camera is opened.

The cameraId field of the config commands sent to the client
was wrongly populated with the local cameraId generated within
vhal. This was causing incorrect mapping of the client cameras.

As the VirtualFakeCamera3 already has the client camera_info,
fixed this issue by sending the config_cmd with the client's
cameraId.

Tracked-On: OAM-103999
Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>