Fix for the crash while releasing the decoder
Observed crash while releasing the decoder since decoding
and releasing operations are happening in two different threads
and accessing decoding memories between these two threads.
And as a result it would be failed to release the decoder
and hence it causes close camera failure.

To fix this issue, introduce a mutex to synchronize the
release and decoder operations.

Change-Id: I8f9f915fc491109808bf9bb1271bac4926641de7
Tracked-On: OAM-104015
Signed-off-by: Muhammad Aksar <muhammad.aksar@intel.com>