Add Gtest support to test Camera-vhal API's
Introduce a property based on which instance id is taken
for socket connection for Gtest purpose. Check for this
property in camera-vhal and if this set then use the
instance id that is set by property for Gtest else use
the default one.

Newly introduced property is ro.boot.container.testid.

Add Unit TestCases for Camera-vhal in Gtest.

Tracked-On: OAM-105831
Signed-off-by: Singh, Sapna1 sapna1.singh@intel.com