Fix for no camera preview in AV1 streaming configuration
When aic is installed with AV1 codec for streaming,
black screen is seen in camera preview.

Codec validation only checks for H264 and H265 which results
in codec type set to H264 even when AV1 is a valid codec type.

Fix the issue by adding AV1 to valid codec type check.

Tracked-On: OAM-110610
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>